# docs/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "GrownupsVet API",
    "description" : "Contrato generado desde código para acceso, perfil, sesiones, mascotas, recuperación, personal con invitaciones y disponibilidad veterinaria. El correo requiere configuración SMTP. La URL de activación del portal contiene un placeholder configurable. Reservas, consulta automática de COMVEZCOL y plantillas Thymeleaf quedan fuera de este incremento.",
    "version" : "0.6.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/users/me/profile/photo" : {
      "get" : {
        "tags" : [ "Foto de perfil propia" ],
        "summary" : "Consultar la foto propia",
        "operationId" : "getCurrentUserProfilePhoto",
        "responses" : {
          "200" : {
            "description" : "Bytes JPEG o PNG procesados",
            "content" : {
              "image/jpeg" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              },
              "image/png" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "La cuenta no tiene foto",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Foto de perfil propia" ],
        "summary" : "Crear o reemplazar la foto propia",
        "description" : "Recibe el campo multipart file. Acepta JPEG/PNG de hasta 2 MiB y almacena una copia procesada sin metadatos.",
        "operationId" : "putCurrentUserProfilePhoto",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "Foto guardada"
          },
          "400" : {
            "description" : "Imagen vacía, inválida o con dimensiones no permitidas",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "413" : {
            "description" : "Archivo superior a 2 MiB",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "El contenido real no es JPEG ni PNG",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Foto de perfil propia" ],
        "summary" : "Eliminar la foto propia",
        "description" : "Es idempotente: devuelve 204 aunque la cuenta no tenga foto.",
        "operationId" : "deleteCurrentUserProfilePhoto",
        "responses" : {
          "204" : {
            "description" : "Foto eliminada o ya inexistente"
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiErrorResponseDTO" : {
        "type" : "object",
        "description" : "Respuesta de error RFC 9457, con código funcional y errores por campo.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "Referencia URI al tipo de problema.",
            "example" : "about:blank"
          },
          "title" : {
            "type" : "string",
            "description" : "Resumen del problema.",
            "example" : "Solicitud incorrecta"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Estado HTTP de la respuesta.",
            "example" : 400,
            "maximum" : 599,
            "minimum" : 400
          },
          "detail" : {
            "type" : "string",
            "description" : "Explicación de lo ocurrido en esta solicitud.",
            "example" : "Revisa los campos indicados."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri-reference",
            "description" : "Referencia URI opcional que identifica esta ocurrencia del problema.",
            "example" : "urn:uuid:7d667530-867b-4c78-a458-0904fb82d574"
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Código funcional estable; el frontend no debe comparar los mensajes.",
            "example" : "INVALID_REQUEST"
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Errores por campo. Es una lista vacía cuando el error no corresponde a campos específicos.",
            "items" : {
              "$ref" : "#/components/schemas/FieldValidationErrorResponseDTO"
            }
          }
        },
        "required" : [ "detail", "errorCode", "fieldErrors", "status", "title", "type" ]
      },
      "FieldValidationErrorResponseDTO" : {
        "type" : "object",
        "description" : "Error asociado a un campo de la solicitud. Nunca contiene el valor recibido.",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "Nombre del campo en el contrato JSON.",
            "example" : "email"
          },
          "code" : {
            "type" : "string",
            "description" : "Código estable que identifica el error del campo.",
            "example" : "TYPE_MISMATCH"
          },
          "message" : {
            "type" : "string",
            "description" : "Explicación legible para la persona usuaria.",
            "example" : "Debe enviarse como texto."
          }
        },
        "required" : [ "code", "field", "message" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT RS256 supplied in the Authorization header. Tokens expire 24 hours after login.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}